refactor (step2) : 저번 리뷰 수정
- Calculator Exception 순서 수정
- valueSplitList[i+1]에 .toIntOrNull()로 에러 핸들링 추가
